model_zoo/ispu_wand/training/requirements.txt
keras<3
tensorflow
matplotlib
numpy
pandas
scipy
seaborn
scikit-learn
pyyaml
schema
